@@ -9,7 +9,7 @@ environment:
   # Opt out of null safety on purpose because dependencies do not support sound null safety yet
   # and flutter drive ignores the --no-sound-null-safety parameter.
   # todo(creativecreatorormaybenot): switch to 2.12.0-0 lower bound.
-  sdk: '>=2.11.0 <3.0.0'
+  sdk: '>=2.12.0 <4.0.0'
 
 dependencies:
   flutter:
